Revert "Merge remote-tracking branch 'upstream/master' into panadapter_v3"
This reverts commit 412d5f6c1cd9fbf270038c77441f993018f49276, reversing
changes made to 3a02ddaf3bbd9598f9d78a8f3cc47237ba231170.

+++ src/applications/gqrx/receiver.cpp
@@ -1010,7 +1010,7 @@ receiver::status receiver::set_af_gain(float gain_db)
     float k;
 
     /* convert dB to factor */
-    k = powf(10.0f, gain_db / 20.0f);
+    k = pow(10.0, gain_db / 20.0);
     //std::cout << "G:" << gain_db << "dB / K:" << k << std::endl;
     audio_gain0->set_k(k);
     audio_gain1->set_k(k);
